refactor(grs): Extract `ReducedMatcherPattern`
Extracting the reduced view on CKEditor's MatcherPattern, as it is used
within GRS. This includes the utility method to merge such patterns,
which again strongly relies on the simplified MatcherPattern.
